frontend: Rename ApplicationItemGroupsList.react.js and List.react.js
