+++ frontend/src/js/components/Applications/ApplicationItemGroupsList.tsx
