+++ frontend/src/js/components/Applications/List.tsx
